Unravel a circular dependency
