+++ utils/ioc_test_framework_utils.py
@@ -1,9 +1,8 @@
 """ Utilities for integrating the device into the IOC test framework """
 from templates.paths import TESTS_TEMPLATE
 from system_paths import IOC_TEST_FRAMEWORK_ROOT
-from shutil import copyfile
 from os import path
-from common_utils import replace_in_file, copy_file
+from file_system_utils import replace_in_file, copy_file
 from device_info_generator import DeviceInfoGenerator
 import logging
